Get rid of TFDeabstraction and move some utilities to TFUtilities.
