+++ lib/SILOptimizer/Transforms/PerformanceInliner.cpp
@@ -26,7 +26,6 @@
 #include "llvm/Support/Debug.h"
 
 // SWIFT_ENABLE_TENSORFLOW
-#include "../lib/SILOptimizer/Mandatory/TFDeabstraction.h"
 #include "../lib/SILOptimizer/Mandatory/TFUtilities.h"
 
 using namespace swift;
@@ -915,8 +914,7 @@ bool SILPerformanceInliner::inlineCallsIntoFunction(SILFunction *Caller) {
                        AI.getSubstitutionMap(), OpenedArchetypesTracker);
 
     // SWIFT_ENABLE_TENSORFLOW
-    if (hasTensorOps &&
-        swift::tf::TFDeabstractionHelper::isSpecialNoInlineCallee(AI, *Callee))
+    if (hasTensorOps && TFC.isSpecialNoInlineCallee(AI, *Callee))
       continue;
 
     // We've already determined we should be able to inline this, so
